Bug 11411 - Bootstrap theme (enhancement) for (right to left)
This patch improves Koha's layout when a right-to-left language is being
displayed by correcting a float set by the right-to-left version of
Bootstrap CSS.

The patch modifies the right-to-left Less file and includes a
now-minified version of the compiled CSS.

To test, install a right-to-left language like ar-Arab. Apply the patch
and view the OPAC under the Bootstrap theme. When logged in, the log out
link is now aligned to the left.

Signed-off-by: Holger Meißner <h.meissner.82@web.de>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>